refac: move user register into separate component
